Read physical maximum for pen interface from HID if possible
The pen interface estimates values for x_phy and y_phy prior to reading
the HID descriptor, using an assumed resolution of 100 points/mm and
the hard-coded logical dimensions. After the actual logical dimensions
are read from the HID descriptor, they are used to calculate the actual
resolution from the estimated physical size. This will produce a value
of 100 (which is normally the case) provided the actual logical dimensions
match the declared dimensions, but can result in incorrect values (e.g.
107 points/mm) if the actual dimensions from the HID descriptor do not
match the declared dimensions.

This patch has us read the physical size from the HID descriptor as well
if it is avaialble. This ensures that we get a correct resolution in
these cases.