Fix potential crash on tablet disconnect
It is possible for an application to trigger a segfault by attempting
to make use of a sysfs command (through a stale fd) after the memory
for the driver data has been freed. To prevent this, we check to
ensure that 'dev_get_drvdata' returns a non-null value before using
it.

WIP: Is there potentially still a window of opportunity where the
driver has a non-null (but free'd) drvdata that we'd dereference?
Also, Is it safe to set the LEDs after they've been "destroyed" but
before the 'wacom' struct has been freed?

Fixes SF bug #255

Signed-off-by: Jason Gerecke <killertofu@gmail.com>

If an application continues to hold a reference to one of the sysfs
nodes while the device is being disconnected, its possible to trigger
a sega