HACK: Properly support the Lenovo Yoga 2
The Lenovo Yoga 2 has a sensor with ID 0xEC, just like the Samsung Slate 7.
Sensors like 0xEC are "generic" in the sense that they can be found in a
number of different devices, so the driver must be able to get specific
hardware details (e.g. of the sensor width, height, and number of supported)
pressure levels.

For some reason, the code in 'wacom_parse_hid' for pre-3.17 kernels does not
seem to be overwriting the pressure (and perhaps the width/height) with
correct values, causing the device to behave incorrectly.

This patch modifies the static definition to work with the Yoga 2, but
obviously this is not an ideal solution. We should fix 'wacom_parse_hid'
instead.