HID: wacom: Correct the retry logic in wacom_set_device_mode
A failure to set the device mode does not trigger a retry as expected.
If the set or get call fails for some reason, they will return a negative
error code which will cause the 'while' condition to evaluate as 'false'.
In addition to fixing the necessary error code condition, we also need
to change the check for `rep_data[1]` being an unexpected value from
being '&&' to '||'. Loosely speaking, we should retry regardless of if
something failed in the call **or** if the hardware reports that it is
in an unexpected mode. More concretely, if the set call fails, the value
of `rep_data[1]` may be left unmodified and cause the check for it to
pass.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>