Prevent the kernel from attempting to sign our module for us
As part of the kernel's `make modules_install` process, it runs the sign-file
program if necessary to try and sign the modules before installation. This
rarely works for most users since the kernel is usually configured to make use
of a distriution signing key which users don't have access to. To provide
users with a signed module, the input-wacom install target runs sign-file on
its own, making use of the MOK keys that exist on the user's machine.

Having the kernel fail at signing the modules itself is fine (nothing bad
happens), but the failure can confuse users who look through the terminal
output and see various "file not found" error messages. This commit disables
the kernel's built-in signing mechanism so that only we run sign-file.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>