Use the kernel's built-in signing framework rather than our own
The kernel tries to sign files as part of 'make modules_install', so we
might as well go along with the flow and let it handle things. Some users
get confused by the error messages that it produces when signing fails
(e.g. because there is no key), so we short-circuit its signing command
when signing is not possible.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>