travis: Add the most recent upstream kernel to the build matrix
Rather than continually updating the script (or neglecting to update it,
as the case may be) to build the most recent kernel, we can use a little
magic to discover the most recent kernel and then perform a build against
it. This should not only ensure that our patches work with upstream but
also ensure that we're using the latest-and-greatest checkpatch.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>