travis: Build with Ubuntu Bionic infrastructure
The Bionic servers come with an updated GCC compiler compared to the
Trusty servers we had been using (7.4.0 versus 4.8.4). This provides
some additional static checks, and is also necessary in order to build
the current Linux 5.2 kernel (which requires the compiler to have
retpoline support).

The GCC update isn't entirely straightforward, however: beginning with
GCC 6 the "-fpie" option is enabled by default. This isn't compatible
with older kernels, so we need to manually disable the option in order
to keep these builds working. We also have to fake some missing headers
for these older kernels. Finally, we also need to update our list of
isystem directories to prevent new spurious warnings.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>