Work around pointer-sign warnings caused by buggy kernel code
Current versions of GCC send up "pointer-sign" warnings in the 3.7 and
earlier trees due to a kernel function and structure not agreeing on
whether the "interface" pointer is signed or not. Rather than silencing
the warning (which is otherwise happy with our code), we can redefine
the kernel's `usb_get_extra_descriptor` macro to perform a cast before
calling the problematic function.

./3.7/wacom_sys.c: In function ‘wacom_retrieve_hid_descriptor’:
./3.7/wacom_sys.c:714:10: error: pointer targets in passing argument 1
 of ‘__usb_get_extra_descriptor’ differ in signedness
 [-Werror=pointer-sign]
  error = usb_get_extra_descriptor(interface, HID_DEVICET_HID, &hid_desc);
          ^~~~~~~~~~~~~~~~~~~~~~~~
In file included from ./include/linux/usb/input.h:12:0,
                 from ./3.7/wacom.h:88,
                 from ./3.7/wacom_sys.c:13:
./include/linux/usb.h:306:5: note: expected ‘char *’ but argument is of
 type ‘unsigned char *’
 int __usb_get_extra_descriptor(char *buffer, unsigned size,
     ^~~~~~~~~~~~~~~~~~~~~~~~~~

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>