HID: wacom: Don't assume battery is connected just because we read a usage
Just because we've processed a usage containing battery level or charge
status doesn't mean that a battery is actually connected. Lets allow the
driver to assume that a battery is not connected if the device reports
the level as 0% and its status as not-charging.

This change is prep for the following unification patch. It ensures that
AES pens (which have this specific state when they go out of prox) will
still be reported with an "unknown" status when they go out of prox.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>