HID: wacom: Defer calculation of resolution until resolution_code is known
The Wacom driver maps the HID_DG_TWIST usage to ABS_Z (rather than ABS_RZ)
for historic reasons. When the code to support twist was introduced in
commit 50066a0, we were careful to write it in such a way that it had
HID calculate the resolution of the twist axis assuming ABS_RZ instead
(so that we would get correct angular behavior). This was broken with
the introduction of commit 08a46b4190, which moved the resolution
calculation to occur *before* the adjustment from ABS_Z to ABS_RZ occured.

This commit moves the calculation of resolution after the point that
we are finished setting things up for its proper use.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Fixes: 08a46b4190 (HID: wacom: Set a default resolution for older tablets)
Cc: stable@vger.kernel.org