HACK: Don't actually remove pen from prox when hung in-prox
It almost certianly should not be the case that sending the pen out of
prox makes jumps worse. However, for the sake of scientific inquiry,
lets give it a shot.

NOTE: Being hung in-prox means that we haven't received an event in a
while. By the very nature of this, it means that the pointer is going
to make a large jump from its last-known location to whatever its new
position is once we start getting events again. Wether we remove the
pen from prox or not, that jump *will* happen.