HID: wacom: Do not allow per-LED brightness values
Our use of the LED interface allows you to set the brightness of a
single LED, but this is not very useful. The status LEDs are set of
related LEDs that should all share the same brightness value. Instead
of requiring users/developers to write to the brightness property of
all LEDs individually, we can have all the LEDs share a common global
brightness so that writing to any one of them will cause all to be
affected.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>