input: wacom_w8001: Resolve name truncation warning
GCC was not able to analyze our assembly of device names well enough to
prove that no truncation would occur. Because of this, a pair of warnings
would be generated at compile-time with W=1. To silence these spurious
warnings, we change how names are assembled to be more straightforward.
We gather all of the information needed to generate a name and pass that
information to a new helper function that does the assembly all in one
go. This makes it much easier (for compilers *and* humans) to analyze
the behavior. The new function still does not truncate, but out of an
abundance of caution, if a truncation /is/ detected we will return an
error code and stop processing.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Reported-by: kernel test robot <lkp@intel.com>
Closes: https://lore.kernel.org/oe-kbuild-all/202503092244.Ets7kUrD-lkp@intel.com/